+++ database/mssql/procedures/read/vam.sql
@@ -3,7 +3,8 @@ DROP PROCEDURE get_vams;
 DROP PROCEDURE get_vams_quadtree;
 DROP PROCEDURE get_vams_quadtree_and_station_id; 
 GO
-CREATE PROCEDURE get_vams @time_start INT, @time_end INT
+CREATE PROCEDURE get_vams @time_start BIGINT,
+@time_end BIGINT
 AS
 select 
   VAM.event_timestamp as [timestamp],
@@ -16,7 +17,10 @@ from
 where 
   VAM.event_timestamp between @time_start and @time_end
 GO
-CREATE PROCEDURE get_vams_station_id @time_start INT, @time_end INT, @in_station_id INT
+CREATE PROCEDURE get_vams_station_id 
+  @time_start BIGINT,
+  @time_end BIGINT,
+  @in_station_id BIGINT
 AS
 select 
   VAM.event_timestamp as [timestamp],
@@ -30,7 +34,11 @@ where
   VAM.event_timestamp between @time_start and @time_end and
   @in_station_id = emitter_station_id
 GO
-CREATE PROCEDURE get_vams_quadtree @time_start INT, @time_end INT, @quadtree_start BIGINT, @quadtree_end BIGINT
+CREATE PROCEDURE get_vams_quadtree 
+  @time_start BIGINT,
+  @time_end BIGINT,
+  @quadtree_start BIGINT,
+  @quadtree_end BIGINT
 AS
 select 
   VAM.event_timestamp as [timestamp],
@@ -44,7 +52,12 @@ where
   VAM.event_timestamp between @time_start and @time_end and
   quadtree between @quadtree_start and @quadtree_end
 GO
-CREATE PROCEDURE get_vams_quadtree_and_station_id @time_start INT, @time_end INT, @quadtree_start BIGINT, @quadtree_end BIGINT, @in_station_id INT
+CREATE PROCEDURE get_vams_quadtree_and_station_id 
+  @time_start BIGINT,
+  @time_end BIGINT,
+  @quadtree_start BIGINT,
+  @quadtree_end BIGINT,
+  @in_station_id BIGINT
 AS
 select 
   VAM.event_timestamp as [timestamp],
